feat(troika-xr): add WristMountedUI component
This facade provides a container for arbitrary global UI, which is
hidden by default but is easily brought up by a simple gesture. A
wristband is added to one of the hand controllers, with an icon
affordance on the inner wrist. When that icon is turned upwards, the
UI is projected from it.