perf(troika-3d-text): move clipping logic from fragment to vertex shader
The clipRect is now evaluated in the vertex shader by shrinking the size
of the glyph quads, rather than adjusting alpha in the fragment shader.
This results in simpler logic running on fewer inputs, and since fully
clipped glyphs are degenerate the fragment shader isn't run at all for
them.