fix(troika-3d-ui): prevent sync of text nodes before flex layout finishes
This prevents a flash of incorrectly-sized text, though it can still
give a flash of old text; that should also be fixed in the future. This
is preferable in the short term because by skipping the initial sync it
frees up the worker thread to more quickly complete the flex layout.
This is particularly noticeable when the text nodes are extremely large.