fix(troika-three-utils): fix program switching when double-deriving materials
Using a single #define TROIKA_DERIVED_MATERIAL is not sufficient now
that we're reusing the same id based on derivation options; this fails
to trigger a program switch if using the same derivation options against
a base material and a derived material based on a similar base material.
E.g. A->C and A->B->C would end up with identical program cache keys so
the renderer wouldn't switch the program. This fixes it by adding
separate #define properties so every derivation level contributes to the
program key.