fix(troika-three-text): fix changing text length in ThreeJS r117+
Fixes #69. This works around an issue introduced in Three r117 where a
given geometry uses the size of any InstancedBufferAttribute as a max
for `instanceCount`, and caches that for the lifetime of the geometry
even if the attribute is replaced with one of a different size.

It's unclear if this is a ThreeJS bug or a if our approach of "resizing"
by replacing the attribute is truly unsupported; the discussion in
https://github.com/mrdoob/three.js/issues/19706 is ambiguous,
https://github.com/mrdoob/three.js/issues/19595 implies it's not
allowed, and https://github.com/mrdoob/three.js/pull/17418 implies it
should be allowed.