fix(troika-three-text): prevent unbound buffer errors when disposing a GlyphsGeometry
Followup fix for #69, reported in react-spring/drei#62

We were reusing the common plane BufferAttributes across GlyphsGeometry
instances, which used to work but as of Three r118 now throws an error
about unbound buffers when one of the geometries is disposed. We now
clone all the base plane attributes, which keeps things isolated at the
expense of redundancy. Thankfully those attributes are usually very
small so it's not a huge amount of duplicated data.