followup: fix artifacts in some glyphs on Windows/Nvidia
For some reason calculating the UV bounds of a glyph within the full texture
within the fragment shader caused strange artifacts on the leftmost glyphs
in the texture. Moving that calculation to the vertex shader and passing
the bounds via varying to be mix()-ed in the fragment shader works properly,
for some reason.