perf(troika-three-text): avoid extra draw call on double sided materials as of Three r130
Three r130 introduced a "feature" which forces DoubleSide transparent materials to be
rendered in two draw calls instead of one; there's no advantage to this for flat planes,
and it can be a significant performance hit. This works around that by making the geometry
two planes back-to-back, always rendered as FrontSide, and emulating the intended sidedness
via drawRange.